refactor: 패키지 구조(api, front 분리) 및 클래스명 변경
