+++ src/test/java/roomescape/ReservationTest.java
@@ -16,7 +16,7 @@ import org.springframework.test.annotation.DirtiesContext;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @DisplayName("예약 테스트")
-public class ReservationApiTest {
+public class ReservationTest {
 
     @BeforeEach
     void init() {
